implemented viewmodel
