+++ src/factory/TierListFactory.java
@@ -1,4 +1,14 @@
 package factory;
 
+import interface_adapter.ViewManagerModel;
+import interface_adapter.tierlist.TierListViewModel;
+import view.TierListView;
+
 public class TierListFactory {
+
+    private TierListFactory() {}
+
+    public static TierListView create(ViewManagerModel viewManagerModel, TierListViewModel tierListViewModel) {
+        return null;
+    }
 }
